openfaas-workload/resnet50/requirements.txt
boto3
onnxruntime
pillow